+++ jwst/ami/ami_analyze.py
@@ -7,6 +7,7 @@ from .find_affine2d_parameters import find_rotation
 from . import instrument_data
 from . import nrm_core
 
+
 from astropy import units as u
 
 
@@ -41,6 +42,8 @@ def apply_LG_plus(input_model, filter_model, oversample, rotation,
         Fringe analysis data
 
     """
+    # Create copy of input_model to avoid overwriting input
+    input_copy = input_model.copy()
 
     # If the input data were taken in full-frame mode, extract a region
     # equivalent to the SUB80 subarray mode to make execution time acceptable.
@@ -52,11 +55,11 @@ def apply_LG_plus(input_model, filter_model, oversample, rotation,
         ysize = 80
         xstop = xstart + xsize - 1
         ystop = ystart + ysize - 1
-        input_model.data = input_model.data[ystart - 1:ystop, xstart - 1:xstop]
-        input_model.dq = input_model.dq[ystart - 1:ystop, xstart - 1:xstop]
-        input_model.err = input_model.err[ystart - 1:ystop, xstart - 1:xstop]
+        input_copy.data = input_copy.data[ystart - 1:ystop, xstart - 1:xstop]
+        input_copy.dq = input_copy.dq[ystart - 1:ystop, xstart - 1:xstop]
+        input_copy.err = input_copy.err[ystart - 1:ystop, xstart - 1:xstop]
 
-    data = input_model.data
+    data = input_copy.data
     dim = data.shape[1]
 
     # Set transformation parameters:
@@ -91,7 +94,7 @@ def apply_LG_plus(input_model, filter_model, oversample, rotation,
     ff_t = nrm_core.FringeFitter(niriss, psf_offset_ff=psf_offset_ff,
                                  oversample=oversample)
 
-    output_model = ff_t.fit_fringes_all(input_model)
+    output_model = ff_t.fit_fringes_all(input_copy)
 
     # Copy header keywords from input to output
     output_model.update(input_model)